refactor(App,Calculator,Constant,Splitter): String.js를 Constant.js로 변경하고 숫자 상수 추가
