+++ src/Calculator.js
@@ -1,17 +1,17 @@
-import { ErrorMessages, EMPTY_STRING } from "./Strings.js";
+import { ErrorMessages, EMPTY_STRING, ZERO } from "./Constant.js";
 
 class Calculator {
   static checkIsPositiveNumber(numStrArray) {
     return numStrArray.map(numStr => {
-      if (numStr === EMPTY_STRING) return 0;
+      if (numStr === EMPTY_STRING) return ZERO;
       const parsedNum = Number(numStr);
-      if (isNaN(parsedNum) || parsedNum <= 0) throw new Error(ErrorMessages.CALCULATOR_CHECK_IS_POSITIVENUMBER);
+      if (isNaN(parsedNum) || parsedNum <= ZERO) throw new Error(ErrorMessages.CALCULATOR_CHECK_IS_POSITIVENUMBER);
       return parsedNum;
     });
   }
 
   static addNumbers(numArray) {
-    return numArray.reduce((sum, num) => sum + num, 0);
+    return numArray.reduce((sum, num) => sum + num, ZERO);
   }
 }
