+++ src/Constant.js
@@ -1,3 +1,7 @@
+const ZERO = 0;
+const RESULT_MAX_LENGTH = 2;
+const CUSTOM_DELIMITER_PREFIX_LENGTH = 2;
+
 const RESULT_PREFIX = "결과 : ";
 const INPUT_MESSAGE = "덧셈할 문자열을 입력해 주세요.\n";
 const EMPTY_STRING = "";
@@ -16,4 +20,4 @@ const ErrorMessages = {
   CALCULATOR_CHECK_IS_POSITIVENUMBER: ERROR_PREFIX + "[ERROR] 양수가 아닌 값은 계산할 수 없습니다."
 };
 
-export { RESULT_PREFIX, INPUT_MESSAGE, EMPTY_STRING, ErrorMessages, DELIMITER_SECTION_START, INPUT_DELIMITER, DefaultDelimiters };
+export { RESULT_PREFIX, INPUT_MESSAGE, EMPTY_STRING, ErrorMessages, DELIMITER_SECTION_START, INPUT_DELIMITER, DefaultDelimiters, ZERO, CUSTOM_DELIMITER_PREFIX_LENGTH, RESULT_MAX_LENGTH };
