+++ src/Splitter.js
@@ -1,12 +1,12 @@
-import { ErrorMessages, DELIMITER_SECTION_START, INPUT_DELIMITER, EMPTY_STRING, DefaultDelimiters } from "./Strings.js";
+import { ErrorMessages, DELIMITER_SECTION_START, INPUT_DELIMITER, EMPTY_STRING, DefaultDelimiters, RESULT_MAX_LENGTH, CUSTOM_DELIMITER_PREFIX_LENGTH } from "./Constant.js";
 
 class Splitter {
   static #DEFAULT_DELIMITERS = [DefaultDelimiters.FIRST_DELIMITER, DefaultDelimiters.SECOND_DELIMITER];
 
   static splitInput(input) {
     const result = input.split(INPUT_DELIMITER);
-    if (result.length > 2) throw new Error(ErrorMessages.SPLITTER_SPLIT_INPUT);
-    if (result.length === 1) result.unshift(EMPTY_STRING);
+    if (result.length > RESULT_MAX_LENGTH) throw new Error(ErrorMessages.SPLITTER_SPLIT_INPUT);
+    if (result.length !== RESULT_MAX_LENGTH) result.unshift(EMPTY_STRING);
     return result;
   }
 
@@ -15,7 +15,7 @@ class Splitter {
 
     if (!delimiterSection.startsWith(DELIMITER_SECTION_START)) throw new Error(ErrorMessages.SPLITTER_ADD_DELIMITER);
 
-    this.#DEFAULT_DELIMITERS.unshift(delimiterSection.slice(2));
+    this.#DEFAULT_DELIMITERS.unshift(delimiterSection.slice(CUSTOM_DELIMITER_PREFIX_LENGTH));
   }
 
   static splitNumberSection(numberSection) {
